fix: refactor and fix upload track component
